+++ src/app/setting/inputs.js
@@ -1,0 +1,16 @@
+const INPUTS = [
+  {
+    name: "name",
+    placeholder: "Nombre",
+    disabled: false,
+    initial: ""
+  },
+  {
+    name: "description",
+    placeholder: "Descripción",
+    disabled: false,
+    initial: ""
+  },
+];
+
+export default INPUTS;
